+++ src/main/java/io/leangen/graphql/generator/mapping/common/EnvironmentInjector.java
@@ -1,6 +1,6 @@
 package io.leangen.graphql.generator.mapping.common;
 
-import graphql.language.Field;
+import graphql.execution.MergedField;
 import io.leangen.geantyref.GenericTypeReflector;
 import io.leangen.geantyref.TypeToken;
 import io.leangen.graphql.annotations.GraphQLEnvironment;
@@ -12,12 +12,10 @@ import io.leangen.graphql.metadata.strategy.value.ValueMapper;
 import java.lang.reflect.AnnotatedType;
 import java.lang.reflect.Parameter;
 import java.lang.reflect.Type;
-import java.util.List;
 import java.util.Set;
 
 public class EnvironmentInjector implements ArgumentInjector {
     
-    private static final Type listOfFields = new TypeToken<List<Field>>(){}.getType();
     private static final Type setOfStrings = new TypeToken<Set<String>>(){}.getType();
     
     @Override
@@ -29,11 +27,8 @@ public class EnvironmentInjector implements ArgumentInjector {
         if (GenericTypeReflector.isSuperType(setOfStrings, params.getType().getType())) {
             return params.getResolutionEnvironment().dataFetchingEnvironment.getSelectionSet().get().keySet();
         }
-        if (Field.class.equals(raw)) {
-            return params.getResolutionEnvironment().fields.get(0);
-        }
-        if (GenericTypeReflector.isSuperType(listOfFields, params.getType().getType())) {
-            return params.getResolutionEnvironment().fields;
+        if (MergedField.class.equals(raw)) {
+            return params.getResolutionEnvironment().dataFetchingEnvironment.getMergedField();
         }
         if (ValueMapper.class.isAssignableFrom(raw)) {
             return params.getResolutionEnvironment().valueMapper;